==> analyses/05-99_inla-collect.Rmd <==
---
title: "2020 excess mortality & voting patterns in CH"
subtitle: "INLA models using GEM"
author: "Radek Panczak"
date: "`r Sys.Date()`"
output:
  rmdformats::robobook:
    code_folding: show
    self_contained: true
    highlight: pygments
editor_options: 
  chunk_output_type: console
knit: (function(inputFile, encoding) {
  rmarkdown::render(inputFile, encoding = encoding, output_dir = "../docs") })
---

<!-- ----------------------------------------------------- -->

```{r setup, include = FALSE}
set.seed(12345)
options(scipen = 999)
options(max.print = "75")

library(pacman)
p_load(tidyverse, magrittr, scales,  
       sf, tmap, MLmetrics, 
       INLA)

import::from("sjmisc", "frq")
```

```{r conflicts, include = FALSE}
tidyverse::tidyverse_conflicts()

conflicted::conflict_scout()
```

```{r knit-setup, include = FALSE}
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())

knitr::opts_chunk$set(cache = FALSE,
                      prompt = FALSE,
                      tidy = FALSE,
                      comment = NA,
                      message = FALSE,
                      warning = FALSE)

knitr::opts_knit$set(width = 75)
```

<!-- ----------------------------------------------------- -->


```{r include=FALSE}
## Mortality data
data = read_rds("data/BfS-closed/monthly_deaths/w_deaths_2015_2020_year_fin.Rds") %>% 
  filter(age != "<40") %>% 
  # strata with double zeroes seem to be crashing
  filter(pop_mid_poi > 0) 

data_male = data %>% 
  filter(sex == "Male") %>% 
  select(-sex) %>% 
  as.data.frame()

data_female = data %>% 
  filter(sex == "Female") %>% 
  select(-sex) %>% 
  as.data.frame()

## Spatial 
kt = read_rds("data/BfS/kt.Rds")
gg = read_rds("data/BfS/gg.Rds")
tg3o = read_rds("data/BfS/tg3o.Rds")
se_alt = read_rds("data/BfS/se_alt.Rds")

## From Ubelix
gem_sex_bym2 = read_rds("results/gem_sex_bym2.Rds")
```

<!-- ----------------------------------------------------- -->

# Model comparison

## DIC {.tabset}

### Males 

```{r echo=FALSE}
dotchart(c(
  gem_sex_bym2$Male$m1$dic$dic,
  gem_sex_bym2$Male$m2$dic$dic,
  gem_sex_bym2$Male$m3$dic$dic,
  gem_sex_bym2$Male$m4$dic$dic
),
labels = c("Model1",
           "Model 2",
           "Model 3",
           "Model 4"))
```

### Females 

```{r echo=FALSE}
dotchart(c(
  gem_sex_bym2$Female$m1$dic$dic,
  gem_sex_bym2$Female$m2$dic$dic,
  gem_sex_bym2$Female$m3$dic$dic,
  gem_sex_bym2$Female$m4$dic$dic
),
labels = c("Model1",
           "Model 2",
           "Model 3",
           "Model 4"))
```

## WAIC {.tabset}

### Males 

```{r echo=FALSE}
dotchart(c(
  gem_sex_bym2$Male$m1$waic$waic,
  gem_sex_bym2$Male$m2$waic$waic,
  gem_sex_bym2$Male$m3$waic$waic,
  gem_sex_bym2$Male$m4$waic$waic
),
labels = c("Model1",
           "Model 2",
           "Model 3",
           "Model 4"))
```

### Females 

```{r echo=FALSE}
dotchart(c(
  gem_sex_bym2$Female$m1$waic$waic,
  gem_sex_bym2$Female$m2$waic$waic,
  gem_sex_bym2$Female$m3$waic$waic,
  gem_sex_bym2$Female$m4$waic$waic
),
labels = c("Model1",
           "Model 2",
           "Model 3",
           "Model 4"))
```

<!-- ----------------------------------------------------- -->

# Best model males

## Model 

```{r include=FALSE}
final_male = gem_sex_bym2$Male
```

```{r}
summary(final_male$m1)
final_male$m1$summary.hyperpar
```

## Year 

```{r echo=FALSE}
final_male_year = final_male$m1$summary.random$id_year %>% 
  mutate(ID = ID + 2014, 
         `0.025quant` = exp(`0.025quant`),
         `0.5quant` = exp(`0.5quant`),
         `0.975quant` = exp(`0.975quant`))

final_male_year %>% 
  ggplot(aes(y = ID)) + 
  geom_pointrange(aes(x = `0.5quant`, xmin = `0.025quant`, xmax = `0.975quant`)) + 
  theme_minimal() + 
  scale_x_log10() + 
  # coord_flip() + 
  ylab("") 
```

## Age 

```{r echo=FALSE}
final_male_age = final_male$m1$summary.random$id_age %>% 
  mutate(`0.025quant` = exp(`0.025quant`),
         `0.5quant` = exp(`0.5quant`),
         `0.975quant` = exp(`0.975quant`))

final_male_age %>% 
  ggplot(aes(y = ID)) + 
  geom_pointrange(aes(x = `0.5quant`, xmin = `0.025quant`, xmax = `0.975quant`)) + 
  theme_minimal() + 
  ylab("Age group") +
  scale_y_continuous(labels = unique(data_male$age)) +
  scale_x_log10()
```

## Observed vs predicted

```{r echo=FALSE}
stopifnot(nrow(final_male$m1$summary.fitted.values) == nrow(data_male))

final_male_fit = bind_cols(
  data_male,
  final_male$m1$summary.fitted.values
) %>% 
  remove_rownames() %>% 
  mutate(split = ifelse(year < 2020, "Pre-pandemic", "Pandemic"))

ggplot(final_male_fit) + 
  geom_abline(intercept = 0, slope = 1, color = "darkorchid") +
  geom_point(aes(x = `0.5quant`, y = observed), alpha = 0.22) + 
  facet_wrap(vars(year)) +
  coord_fixed() +
  xlab("predicted") + 
  theme_minimal()
```

## Examples

```{r include=FALSE}
ploterr = function(data, community){
  data %>% 
    filter(GMDNAME == community) %>% 
    ggplot(aes(x = year, fill = age, group = age)) + 
    geom_ribbon(aes(ymin = `0.025quant`, ymax = `0.975quant`), alpha = 0.1) +
    geom_line(aes(y = `0.5quant`, color = age)) + 
    geom_point(aes(y = observed, color = age), alpha = 0.5) + 
    # facet_wrap(vars(split), scales = "free_x") +
    xlab("") + ylab("predicted/observed") + ggtitle(community) + 
    theme_minimal()
}
```

### Lugano 

```{r echo=FALSE}
ploterr(final_male_fit, "Lugano")
```

## Errors examples 

```{r echo=FALSE}
errors = final_male_fit %>% 
  filter(year < 2020) %>% 
  group_by(GMDNAME) %>% 
  summarize(pred = sum(as.integer(round(`0.5quant`))),
            deaths = sum(deaths)) %>% 
  ungroup() %>% 
  mutate(diff = deaths - pred,
         perc = ((deaths - pred) / deaths) * 100) %>% 
  # zero predicted
  mutate(perc = if_else(pred == 0 & deaths > 0, diff * 100, perc)) %>% 
  # zero deaths
  mutate(perc = if_else(deaths == 0 & pred > 0, diff * 100, perc))
```

### Zero observed 2015-19

```{r echo=FALSE}
ploterr(final_male_fit, "Schwende-Rüte")
```

### Zero predicted 2015-19

```{r echo=FALSE}
ploterr(final_male_fit, "Saint-Brais")
```

### Large(ish) absolute error

```{r echo=FALSE}
ploterr(final_male_fit, "Appenzell")
```

## Aggregating all observed and pred over years

```{r include=FALSE}
final_male_fit_agg = final_male_fit %>% 
  group_by(year) %>% 
  summarise(observed = sum(observed),
            # population = sum(population),
            predicted = as.integer(sum(`0.5quant`))) %>% 
  ungroup() %>% 
  mutate(diff = observed - predicted)
```

```{r echo=FALSE}
final_male_fit_agg
```

```{r echo=FALSE}
final_male_fit_agg %>% 
  ggplot(aes(x = year)) + 
  geom_line(aes(y = predicted)) + 
  geom_line(aes(y = observed), col = "darkorchid", alpha = 0.5) + 
  # facet_wrap(vars(split), scales = "free_x") +
  xlab("") + ylab("predicted/observed") +
  theme_minimal()
```

## Region effect 

```{r eval=FALSE, include=FALSE}
# misc checks

length(data_male$id_space)
length(unique(data_male$id_space))

length(tg3o$GMDNR)
length(unique(tg3o$GMDNR))

length(final_male$m1$summary.random$id_space$ID)
length(unique(final_male$m1$summary.random$id_space$ID))

summary(data_male$id_space)
summary(final_male$m1$summary.random$id_space$ID)
```

```{r include=FALSE}
stopifnot(length(tg3o$GMDNR) * 2 == length(final_male$m1$summary.random$id_space$ID))
```

```{r include=FALSE}
final_male_space =
  tg3o %>%
  select(GMDNR) %>% 
  left_join(data_male %>% select(GMDNR, id_space) %>% distinct()) %>%
  left_join(final_male$m1$summary.random$id_space %>%
              mutate(s1_l = exp(`0.025quant`),
                     s1_m = exp(`0.5quant`),
                     s1_u = exp(`0.975quant`)) %>% 
              select(ID, starts_with("s1")),
            by = c("id_space" = "ID")) %>% 
  mutate(s1_m = cut(s1_m, 
                    breaks = c(0, 0.64, 0.86, 0.95, 1.05, 1.16, 1.57, 10),
                    right = FALSE))
```

```{r}
frq(final_male_space$s1_m)
```

```{r echo=FALSE}
tm_shape(se_alt, is.master = FALSE) +
  tm_fill(col = "#a7cdf2") + 
  tm_shape(final_male_space, is.master = TRUE) +
  tm_fill("s1_m", palette = "-PuOr", 
          title = "Region effect") + 
  tm_shape(kt) +
  tm_borders(col = "grey60")
```

```{r include=FALSE}
final_male_space =
  tg3o %>%
  select(GMDNR) %>% 
  left_join(data_male %>% 
              filter(year == 2019) %>% 
              group_by(GMDNR, id_space) %>% 
              summarise(deaths = sum(deaths)) %>% 
              ungroup()) %>%
  left_join(final_male$m1$summary.random$id_space %>%
              mutate(s1_l = exp(`0.025quant`),
                     s1_m = exp(`0.5quant`),
                     s1_u = exp(`0.975quant`)) %>% 
              select(ID, starts_with("s1")),
            by = c("id_space" = "ID")) %>% 
  mutate(s1_m = cut(s1_m, 
                    breaks = c(0, 0.64, 0.86, 0.95, 1.05, 1.16, 1.57, 10),
                    right = FALSE))
```

```{r echo=FALSE}
tm_shape(se_alt, is.master = FALSE) +
  tm_fill(col = "#a7cdf2") + 
  tm_shape(kt, is.master = TRUE) +
  tm_borders() +  
  tm_shape(final_male_space) +
  tm_symbols(size = "deaths", col = "s1_m",
             shape = 15, scale = 2, perceptual = TRUE, 
             legend.size.show = FALSE, 
             palette = "-PuOr", 
             title.size = "Population", 
             title.col = "Region effect") + 
  tm_legend(position = c("LEFT", "TOP"))
```

## Performance 

```{r}
marg.variance = inla.tmarginal(function(x) 1/x,
                               final_male$m1$marginals.hyperpar$"Precision for id_space")
inla.zmarginal(marg.variance)
```

```{r}
table(final_male$m1$cpo$failure > 0)
sum(-log(final_male$m1$cpo$cpo), na.rm = TRUE)
```

# ~~Best~~ First model females

## Model 

```{r include=FALSE}
final_female = gem_sex_bym2$Female
```

```{r}
summary(final_female$m1)
final_female$m1$summary.hyperpar
```

## Year 

```{r echo=FALSE}
final_female_year = final_female$m1$summary.random$id_year %>% 
  mutate(ID = ID + 2014, 
         `0.025quant` = exp(`0.025quant`),
         `0.5quant` = exp(`0.5quant`),
         `0.975quant` = exp(`0.975quant`))

final_female_year %>% 
  ggplot(aes(y = ID)) + 
  geom_pointrange(aes(x = `0.5quant`, xmin = `0.025quant`, xmax = `0.975quant`)) + 
  theme_minimal() + 
  scale_x_log10() + 
  # coord_flip() + 
  ylab("") 
```

## Age 

```{r echo=FALSE}
final_female_age = final_female$m1$summary.random$id_age %>% 
  mutate(`0.025quant` = exp(`0.025quant`),
         `0.5quant` = exp(`0.5quant`),
         `0.975quant` = exp(`0.975quant`))

final_female_age %>% 
  ggplot(aes(y = ID)) + 
  geom_pointrange(aes(x = `0.5quant`, xmin = `0.025quant`, xmax = `0.975quant`)) + 
  theme_minimal() + 
  ylab("Age group") +
  scale_y_continuous(labels = unique(data_female$age)) +
  scale_x_log10()
```

## Observed vs predicted

```{r echo=FALSE}
stopifnot(nrow(final_female$m1$summary.fitted.values) == nrow(data_female))

final_female_fit = bind_cols(
  data_female,
  final_female$m1$summary.fitted.values
) %>% 
  remove_rownames() %>% 
  mutate(split = ifelse(year < 2020, "Pre-pandemic", "Pandemic"))

ggplot(final_female_fit) + 
  geom_abline(intercept = 0, slope = 1, color = "darkorchid") +
  geom_point(aes(x = `0.5quant`, y = observed), alpha = 0.22) + 
  facet_wrap(vars(year)) +
  coord_fixed() +
  xlab("predicted") + 
  theme_minimal()
```

## Examples

### Lugano 

```{r echo=FALSE}
ploterr(final_female_fit, "Lugano")
```

<!-- ## Errors  -->

```{r include=FALSE}
errors = final_female_fit %>% 
  filter(year < 2020) %>% 
  group_by(GMDNAME) %>% 
  summarize(pred = sum(as.integer(round(`0.5quant`))),
            deaths = sum(deaths)) %>% 
  ungroup() %>% 
  mutate(diff = deaths - pred,
         perc = ((deaths - pred) / deaths) * 100) %>% 
  # zero predicted
  mutate(perc = if_else(pred == 0 & deaths > 0, diff * 100, perc)) %>% 
  # zero deaths
  mutate(perc = if_else(deaths == 0 & pred > 0, diff * 100, perc))
```

## Aggregating all observed and pred over time

```{r include=FALSE}
final_female_fit_agg = final_female_fit %>% 
  group_by(year) %>% 
  summarise(observed = sum(observed),
            # population = sum(population),
            predicted = as.integer(sum(`0.5quant`))) %>% 
  ungroup() %>% 
  mutate(diff = observed - predicted)
```

```{r echo=FALSE}
final_female_fit_agg
```

```{r echo=FALSE}
final_female_fit_agg %>% 
  ggplot(aes(x = year)) + 
  geom_line(aes(y = predicted)) + 
  geom_line(aes(y = observed), col = "darkorchid", alpha = 0.5) + 
  # facet_wrap(vars(split), scales = "free_x") +
  xlab("") + ylab("predicted/observed") +
  theme_minimal()
```

## Region effect 

```{r eval=FALSE, include=FALSE}
# misc checks

length(data_female$id_space)
length(unique(data_female$id_space))

length(tg3o$GMDNR)
length(unique(tg3o$GMDNR))

length(final_female$m1$summary.random$id_space$ID)
length(unique(final_female$m1$summary.random$id_space$ID))

summary(data_female$id_space)
summary(final_female$m1$summary.random$id_space$ID)
```

```{r include=FALSE}
stopifnot(length(tg3o$GMDNR) * 2 == length(final_female$m1$summary.random$id_space$ID))
```

```{r include=FALSE}
final_female_space =
  tg3o %>%
  select(GMDNR) %>% 
  left_join(data_female %>% select(GMDNR, id_space) %>% distinct()) %>%
  left_join(final_female$m1$summary.random$id_space %>%
              mutate(s1_l = exp(`0.025quant`),
                     s1_m = exp(`0.5quant`),
                     s1_u = exp(`0.975quant`)) %>% 
              select(ID, starts_with("s1")),
            by = c("id_space" = "ID")) %>% 
  mutate(s1_m = cut(s1_m, 
                    breaks = c(0, 0.64, 0.86, 0.95, 1.05, 1.16, 1.57, 10),
                    right = FALSE))

frq(final_female_space$s1_m)
```

```{r echo=FALSE}
tm_shape(se_alt, is.master = FALSE) +
  tm_fill(col = "#a7cdf2") + 
  tm_shape(final_female_space, is.master = TRUE) +
  tm_fill("s1_m", palette = "-PuOr", 
          title = "Region effect") + 
  tm_shape(kt) +
  tm_borders(col = "grey60")
```

```{r include=FALSE}
final_female_space =
  tg3o %>%
  select(GMDNR) %>% 
  left_join(data_female %>% 
              filter(year == 2019) %>% 
              group_by(GMDNR, id_space) %>% 
              summarise(deaths = sum(deaths)) %>% 
              ungroup()) %>%
  left_join(final_female$m1$summary.random$id_space %>%
              mutate(s1_l = exp(`0.025quant`),
                     s1_m = exp(`0.5quant`),
                     s1_u = exp(`0.975quant`)) %>% 
              select(ID, starts_with("s1")),
            by = c("id_space" = "ID")) %>% 
  mutate(s1_m = cut(s1_m, 
                    breaks = c(0, 0.64, 0.86, 0.95, 1.05, 1.16, 1.57, 10),
                    right = FALSE))
```

```{r echo=FALSE}
tm_shape(se_alt, is.master = FALSE) +
  tm_fill(col = "#a7cdf2") + 
  tm_shape(kt, is.master = TRUE) +
  tm_borders() +  
  tm_shape(final_female_space) +
  tm_symbols(size = "deaths", col = "s1_m",
             shape = 15, scale = 2, perceptual = TRUE, 
             legend.size.show = FALSE, 
             palette = "-PuOr", 
             title.size = "Population", 
             title.col = "Region effect") + 
  tm_legend(position = c("LEFT", "TOP"))
```

## Performance 

```{r}
marg.variance = inla.tmarginal(function(x) 1/x,
                               final_female$m1$marginals.hyperpar$"Precision for id_space")
inla.zmarginal(marg.variance)
```

```{r}
table(final_female$m1$cpo$failure > 0)
sum(-log(final_female$m1$cpo$cpo), na.rm = TRUE)
```
