==> analyses/07-03_inla-2stage.Rmd <==
---
title: "2020 excess mortality & voting patterns in CH"
subtitle: "INLA 2 stage models using GEM"
author: "Radek Panczak"
date: "`r Sys.Date()`"
output:
  rmdformats::robobook:
    code_folding: show
    self_contained: true
    highlight: pygments
editor_options: 
  chunk_output_type: console
knit: (function(inputFile, encoding) {
  rmarkdown::render(inputFile, encoding = encoding, output_dir = "../docs") })
---

<!-- ----------------------------------------------------- -->

```{r setup, include = FALSE}
set.seed(12345)
options(scipen = 999)
options(max.print = "75")

library(pacman)
p_load(tidyverse, magrittr, scales,  
       sf, tmap, 
       INLA)

import::from("sjmisc", "frq")
```

```{r conflicts, include = FALSE}
tidyverse::tidyverse_conflicts()

conflicted::conflict_scout()
```

```{r knit-setup, include = FALSE}
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())

knitr::opts_chunk$set(cache = FALSE,
                      prompt = FALSE,
                      tidy = FALSE,
                      comment = NA,
                      message = FALSE,
                      warning = FALSE)

knitr::opts_knit$set(width = 75)
```

<!-- ----------------------------------------------------- -->

```{r include=FALSE}
## Mortality data
data = read_rds("data/BfS-closed/monthly_deaths/w_deaths_2015_2020_year_fin.Rds") %>% 
  filter(age != "<40") %>% 
  # strata with double zeroes seem to be crashing
  filter(pop_mid_poi > 0) %>% 
  # only data till 2019
  filter(year < 2020)

data_male = data %>% 
  filter(sex == "Male") %>% 
  select(-sex) %>% 
  as.data.frame()

data_female = data %>% 
  filter(sex == "Female") %>% 
  select(-sex) %>% 
  as.data.frame()

## Spatial 
kt = read_rds("data/BfS/kt.Rds")
gg = read_rds("data/BfS/gg.Rds")
tg3o = read_rds("data/BfS/tg3o.Rds")
se_alt = read_rds("data/BfS/se_alt.Rds")
```

<!-- ----------------------------------------------------- -->

# Stage 1 

Model fitted using 2015-2019 data. Males only.  

```{r}
gem_bym2_19 = read_rds("results_inla/gem_bym2_19.Rds")
gem_bym2_19_kt = read_rds("results_inla/gem_bym2_19_kt.Rds")
```

## Canton {.tabset}

### DIC

```{r echo=FALSE}
dotchart(c(gem_bym2_19$Male$dic$dic, gem_bym2_19_kt$Male$dic$dic), 
         labels = c("-canton", "+canton"),
         cex = 0.6, xlab = "DIC")
```

### WAIC 

```{r echo=FALSE}
dotchart(c(gem_bym2_19$Male$waic$waic, gem_bym2_19_kt$Male$waic$waic), 
         labels = c("-canton", "+canton"),
         cex = 0.6, xlab = "WAIC")
```

## Model 

```{r}
summary(gem_bym2_19_kt$Male)
```

## Calculate the lambda 

```{r}
draws = inla.posterior.sample(50, gem_bym2_19$Male)

rate_draws = lapply(draws, function(i)
  exp(i$latent[grep("Predictor", rownames(i$latent))]))

stopifnot(length(rate_draws[[1]]) == dim(data_male)[1])

rate_draws_med = array(unlist(rate_draws), dim = c(dim(data_male)[1], 50))
stopifnot(dim(rate_draws_med)[1] == dim(data_male)[1])

data_male_draws = cbind(data_male, as.data.frame(rate_draws_med))

data_male_draws_median = data_male_draws %>% 
  group_by(age, GMDNR) %>% 
  summarise_at(vars(V1:V50), mean)  
```

## Merge with the 2020 data

```{r eval=FALSE, include=FALSE}
# strata without join to orig data
# are the ones with excluded years (pop == 0)
anti_join(read_rds("data/BfS-closed/monthly_deaths/w_deaths_2015_2020_year_fin.Rds") %>% 
            filter(age != "<40") %>% 
            filter(year < 2020) %>% 
            filter(sex == "Male"),
          data_male_draws_median) %>% 
  summarise_at(c("pop_mid_poi"), max, na.rm = TRUE)

# example
data_male %>% 
  select(year, age, GMDNR, GMDNAME, observed, pop_mid_poi) %>%  
  filter(age == "80+" & GMDNR == 389)

read_rds("data/BfS-closed/monthly_deaths/w_deaths_2015_2020_year_fin.Rds") %>% 
  filter(sex == "Male") %>% 
  select(year, age, GMDNR, GMDNAME, observed, pop_mid_poi) %>%  
  filter(age == "80+" & GMDNR == 389)
```

```{r}
data_2020 = left_join(
  read_rds("data/BfS-closed/monthly_deaths/w_deaths_2015_2020_year_fin.Rds") %>% 
    filter(year == 2020) %>% 
    filter(age != "<40") %>% 
    filter(sex == "Male") %>% 
    mutate(deaths = observed) %>% 
    select(-(ARGRNR:ARNAME), -observed, - year), 
  data_male_draws_median)
```

## Create 50 datafiles 

1 for each of the samples we extracted, to be used as data in stage 2  

```{r}
for (i in 1:50){
  sample_males = data_2020[, c(1:17, (i+17))]
  
  if (i < 10) {
    suffix = paste0("0", i)
  } else {
    suffix = paste0(i)
  } 
  
  write_rds(sample_males, 
            file = paste("samples/sample_males_", suffix, ".Rds", sep = ""))
}
```

<!-- ----------------------------------------------------- -->

# Stage 2

## Data preps

Sample created above gets merged with voting dataset  

```{r}
data_2020 = read_rds("samples/sample_males_01.Rds") 
```

Communities without `exp`

```{r}
data_2020 %>% 
  filter(is.na(.[[18]])) %>% 
  select(GMDNAME, sex, age, deaths, pop_mid_poi)
```

```{r}
data_2020 = data_2020 %>% 
  filter(!is.na(.[[18]]))
```

Voting data 

```{r}
covid_jun = read_rds("data/voting/covid_jun.Rds") %>% 
  select(GMDNR, vote_yes, jaStimmenInProzent) %>% 
  distinct()
```

Quintiles of votes:  

```{r echo=FALSE}
frq(covid_jun, vote_yes)
```

Range of 'yes' votes within quintile:  

```{r echo=FALSE}
covid_jun %>% 
  group_by(vote_yes) %>% 
  summarise(min = min(jaStimmenInProzent), max = max(jaStimmenInProzent))
```

## Set the expected column 

Using the index we want to run

```{r}
data_2020 = data_2020 %>% 
  mutate(exp = as.integer(pop_mid_poi * .[[18]])) %>%  # the 18th column is the lambda
  select(-18)
```

Missing check  

```{r}
sapply(data_2020, function(x) sum(is.na(x)))
```

Zeros have to handled 

```{r}
frq(data_2020, exp == 0)

data_2020 = data_2020 %>% 
  mutate(exp = ifelse(exp == 0, 0.0000001, exp))
```

Merge with data and clean communities without vote info and with zero pops  

```{r}
data_2020 = left_join(data_2020, covid_jun) %>% 
  # filter few communities without voting info
  filter(!is.na(vote_yes))
```

```{r eval=FALSE, include=FALSE}
frq(data_2020, vote_yes)
sapply(data_2020, function(x) sum(is.na(x)))
```

## Model run

The stage 2 models are run 50 times (for each sex) using the samples extracted in stage 1.  

```{r}
hyper.iid = list(theta = list(prior = "pc.prec", param = c(1, 0.01)))

hyper.bym2 = list(theta1 = list("PCprior", c(1, 0.01)), 
                  theta2 = list("PCprior", c(0.5, 0.5)))

threads = parallel::detectCores()
```

```{r}
formula_cru = deaths ~ 1 + offset(log(exp)) + 
  f(id_age, model = "iid", hyper = hyper.iid, constr = TRUE) + 
  f(id_space, model = "bym2", graph = "data/nb/gg_wm_q.adj", scale.model = TRUE, hyper = hyper.bym2) +
  as.factor(vote_yes) 

model_cru = inla(formula = formula_cru,
                 data = data_2020,
                 family = "Poisson",
                 control.compute = list(config = TRUE, dic = TRUE),
                 quantiles = c(0.025, 0.5, 0.975),
                 num.threads = threads,
                 safe = TRUE,
                 verbose = FALSE)
```

## Results 

```{r}
summary(model_cru)
```

```{r}
exp(model_cru$summary.fixed)
```

## Draws 

```{r}
draws_2 = inla.posterior.sample(100, model_cru)

rate_draws_2 = lapply(draws_2, function(i)
  exp(i$latent[grep("Predictor", rownames(i$latent))]))

# length(rate_draws_2[[1]]) == dim(data_2020)[1]

rate_draws_2_med = array(unlist(rate_draws_2), dim = c(dim(data_2020)[1], 100))
# dim(rate_draws_2_med)

data_2020_male_draws = cbind(data_2020, as.data.frame(rate_draws_2_med))

#Once the batch is complete there will be 5000 samples (50 from stage 1 * 100 from stage 2) which are used to extract the various measure of excess mortality used in the paper. Plus the rate ratios for the covariates.
```

```{r}
sapply(data_2020_male_draws, function(x) sum(is.na(x)))

summary(data_2020_male_draws$V1)
```



```{r eval=FALSE, include=FALSE}
formula_adj = deaths ~ 1 + offset(log(exp)) + 
  f(id_age, model = "iid", hyper = hyper.iid, constr = TRUE) + 
  f(id_space, model = "bym2", graph = "data/nb/gg_wm_q.adj", scale.model = TRUE, hyper = hyper.bym2) + 
  as.factor(vote_yes) + 
  as.factor(border) +
  as.factor(median_ssep3_q) +
  as.factor(r_urban2) +
  as.factor(r_lang)

model_adj = inla(formula = formula_adj,
                 data = data_2020,
                 family = "Poisson",
                 control.compute = list(config = TRUE, dic = TRUE),
                 quantiles = c(0.025, 0.5, 0.975),
                 num.threads = threads,
                 safe = TRUE,
                 verbose = FALSE)

summary(model_adj)
exp(model_adj$summary.fixed)

draws_3 = inla.posterior.sample(100, model_adj)

```
