==> analyses/08_redistribute.Rmd <==
---
title: "2020 excess mortality & voting patterns in CH"
subtitle: "Redistributed cantonal deaths"
author: "Radek Panczak"
date: "`r Sys.Date()`"
output:
  rmdformats::robobook:
    code_folding: show
    self_contained: true
    highlight: pygments
editor_options: 
  chunk_output_type: console
knit: (function(inputFile, encoding) {
  rmarkdown::render(inputFile, encoding = encoding, output_dir = "../docs") })
---

<!-- ----------------------------------------------------- -->

```{r setup, include = FALSE}
set.seed(12345)
options(scipen = 999)
options(max.print = "75")

library(pacman)
p_load(tidyverse, magrittr, skimr, scales,
       hrbrthemes, 
       sf, tmap,
       osfr)

import::from("sjmisc", "frq")
```

```{r conflicts, include = FALSE}
tidyverse::tidyverse_conflicts()

conflicted::conflict_scout()
```

```{r knit-setup, include = FALSE}
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())

knitr::opts_chunk$set(cache = FALSE,
                      prompt = FALSE,
                      tidy = FALSE,
                      comment = NA,
                      message = FALSE,
                      warning = FALSE)

knitr::opts_knit$set(width = 75)
```

<!-- ----------------------------------------------------- -->

# Data 

## Canton 

Expected deaths by age X sex X canton in 2020 from [Riou *et al.*](https://www.medrxiv.org/content/10.1101/2022.08.05.22278458v1).   

```{r include=FALSE}
exp_deaths_2020_kt = read_rds("data/canton/exp_deaths_2020_kt.rds") %>% select(-year)
```

Contains iterations of diff results  

```{r}
max(exp_deaths_2020_kt$it)
```

Example of one iteration  

```{r echo=FALSE}
exp_deaths_2020_kt %>% 
  filter(it == 1) %>% 
  skim()
```

```{r eval=FALSE, include=FALSE}
exp_deaths_2020_kt %>% 
  filter(it == 1) %>% 
  filter(age_group != "<40") %>% 
  frq(exp_deaths > observed2)
```

## Municipality 

Data prepared in `02.Rmd`. Collapsing age groups and renaming variables to match cantonal estimates.  

```{r}
w_deaths_2020_year_fin = read_rds("data/BfS-closed/monthly_deaths/w_deaths_2015_2020_year_fin.Rds") %>% 
  filter(year == 2020) %>% select(-year) %>% 
  rename(age_group = age,
         canton = KTNAME) %>% 
  mutate(age_group = if_else(age_group == "40-49", "40-59", age_group),
         age_group = if_else(age_group == "50-59", "40-59", age_group)) %>% 
  group_by(canton, GMDNR, age_group, sex) %>% 
  summarise(observed = sum(observed),
            pop_mid_poi = sum(pop_mid_poi)) %>% 
  ungroup()
```

```{r echo=FALSE}
w_deaths_2020_year_fin %>% 
  skim()
```

## Spatial 

```{r}
gg = read_rds("data/BfS/gg.Rds")
```

## Downscale function 

@jriou method to downscale cantonal deaths to municipality level.  

```{r}
source("R/downscale_year.R")
```

<!-- ----------------------------------------------------- -->

# Downscale 

```{r}
exp_deaths_2020_year = downscale_year(exp_deaths_2020_kt, w_deaths_2020_year_fin) %>% 
  rename(
    munici_pop = pop_mid_poi,
    munici_observed = observed,
    cant_exp_deaths = exp_deaths,
    cant_observed = observed2) 
```

Again, result contains iterations of diff results  

```{r}
max(exp_deaths_2020_year$it)
```

Example of one iteration  

```{r echo=FALSE}
exp_deaths_2020_year %>% 
  filter(it == 1) %>% 
  skim()
```

```{r include=FALSE}
rm(exp_deaths_2020_kt, w_deaths_2020_year_fin, downscale_year); gc()
```

<!-- ----------------------------------------------------- -->

# Preps

## Aggregates 

Iterations collapsed for each municipality strata.  

```{r}
exp_deaths_2020_year_agg = exp_deaths_2020_year %>% 
  group_by(canton, GMDNR, age_group, sex) %>% 
  summarise(
    # observed
    munici_observed = first(munici_observed),
    munici_pop = first(munici_pop),
    # modelled x1000
    # expected
    munici_expected_lo = quantile(munici_exp_deaths, 0.025),
    munici_expected_med = quantile(munici_exp_deaths, 0.5),
    munici_expected_up = quantile(munici_exp_deaths, 0.975),
    # excess
    munici_excess_lo = quantile(munici_excess, 0.025),
    munici_excess_med = quantile(munici_excess, 0.5),
    munici_excess_up = quantile(munici_excess, 0.975)
  ) %>% 
  ungroup()
```

Even simpler aggregation to municipality level, ignoring age & sex strata.  

```{r}
exp_deaths_2020_year_gem = exp_deaths_2020_year_agg %>% 
  group_by(canton, GMDNR) %>% 
  summarise(
    # community level totals
    munici_pop = sum(munici_pop),
    munici_observed = sum(munici_observed),
    munici_expected_med = sum(munici_expected_med),
    munici_excess_med = sum(munici_excess_med)
  ) %>% 
  ungroup() %>% 
  # ratio and per pop
  mutate(munici_excess_rat = munici_excess_med / munici_expected_med,
         munici_excess_pop =  (munici_excess_med / munici_pop) * 1000) 
```

Special situation arises in this aggregation when calculating a *ratio of excess deaths to observed* deaths. There are communities where expected number of deaths is zero. That leads to ratio being `Inf` in cases where excess is > 0 and to `NaN` where excess is (also) == 0.  

```{r echo=FALSE}
# frq(exp_deaths_2020_year_gem, munici_expected_med == 0 & munici_excess_med != 0)
# frq(exp_deaths_2020_year_gem, munici_expected_med == 0 & munici_excess_med == 0)
frq(exp_deaths_2020_year_gem, munici_expected_med == 0)
```

These communities" ratios were recoded to `NA`.  

```{r include=FALSE}
exp_deaths_2020_year_gem %<>% 
  mutate(munici_excess_rat = ifelse(munici_expected_med == 0, NA, munici_excess_rat))
```

## Bring covariates 

```{r}
covariates = read_rds("data/BfS-closed/monthly_deaths/w_deaths_2015_2020_year_fin.Rds") %>% 
              select(GMDNR, GMDNAME, border, 
                     median_ssep3_q, r_urban1, r_urban2, r_lang,
                     gemtyp_9, gemtyp_9_lab, gemtyp_25, gemtyp_25_lab) %>% 
              distinct()

exp_deaths_2020_year %<>% 
  left_join(covariates)

exp_deaths_2020_year_agg %<>% 
  left_join(covariates)

exp_deaths_2020_year_gem %<>% 
  left_join(covariates)
```

## Bring voting  

```{r}
covid_jun = read_rds("data/voting/covid_jun.Rds") %>% 
  select(GMDNR, jaStimmenInProzent, vote_yes) %>% 
  rename(vote_yes_jun_perc = jaStimmenInProzent,
         vote_yes_jun_cat = vote_yes)

covid_nov = read_rds("data/voting/covid_nov.Rds") %>% 
  select(GMDNR, jaStimmenInProzent, vote_yes) %>% 
  rename(vote_yes_nov_perc = jaStimmenInProzent,
         vote_yes_nov_cat = vote_yes)
```

Note: there are few small communities with no voting info - **these were excluded**.  

```{r echo=FALSE}
exp_deaths_2020_year_gem %>%
  anti_join(covid_jun) %>% 
  select(canton, GMDNAME)
```

```{r include=FALSE}
exp_deaths_2020_year_agg %<>% 
  left_join(covid_jun) %>% 
  filter(!is.na(vote_yes_jun_cat))

exp_deaths_2020_year_gem %<>% 
  left_join(covid_jun) %>% 
  filter(!is.na(vote_yes_jun_cat))

exp_deaths_2020_year_agg %<>% 
  left_join(covid_nov) %>% 
  filter(!is.na(vote_yes_nov_cat)) %>% 
  relocate(GMDNAME, .after = GMDNR)

exp_deaths_2020_year_gem %<>% 
  left_join(covid_nov) %>% 
  filter(!is.na(vote_yes_nov_cat)) %>% 
  relocate(GMDNAME, .after = GMDNR)

exp_deaths_2020_year %<>% 
  left_join(covid_jun) %>% 
  # filter(!is.na(vote_yes_jun_cat)) %>%
  left_join(covid_nov) %>% 
  # filter(!is.na(vote_yes_nov_cat)) %>% 
  relocate(GMDNAME, .after = GMDNR) %>% 
  # adding INLA id vars
  left_join(gg %>% 
              st_drop_geometry() %>% 
              select(GMDNR, id_space)) %>% 
  relocate(id_space, .before = GMDNR) %>% 
  mutate(id_age = as.integer(as.factor(age_group))) %>% 
  relocate(id_space, .after = age_group)

write_rds(exp_deaths_2020_year, "results/exp_deaths_2020_year.Rds")
write_rds(exp_deaths_2020_year_agg, "results/exp_deaths_2020_year_agg.Rds")
write_rds(exp_deaths_2020_year_gem, "results/exp_deaths_2020_year_gem.Rds")
```

```{r include=FALSE}
osf_auth(readr::read_file("secrets/osf_pat.txt"))

project = osf_retrieve_node("gcxt7")
project

project %>%
  osf_ls_nodes() 

upload_location = project %>%
  osf_ls_nodes("Results") 

zip(zipfile = "results/exp_deaths_2020_year.zip", 
    files = "results/exp_deaths_2020_year.Rds")

osf_upload(upload_location, 
           path = "results/exp_deaths_2020_year.zip", 
           recurse = TRUE, 
           conflicts = "overwrite",
           verbose = TRUE)

unlink("results/exp_deaths_2020_year.zip")
```
