==> analyses/09_eda-sir.Rmd <==
---
title: "2020 excess mortality & voting patterns in CH"
subtitle: "Redistributed cantonal deaths"
author: "Radek Panczak"
date: "`r Sys.Date()`"
output:
  rmdformats::robobook:
    code_folding: show
    self_contained: true
    highlight: pygments
editor_options: 
  chunk_output_type: console
knit: (function(inputFile, encoding) {
  rmarkdown::render(inputFile, encoding = encoding, output_dir = "../docs") })
---

<!-- ----------------------------------------------------- -->

```{r setup, include = FALSE}
set.seed(12345)
options(scipen = 999)
options(max.print = "75")

library(pacman)
p_load(tidyverse, magrittr, skimr, scales,
       hrbrthemes, 
       sf, tmap)

import::from("sjmisc", "frq")
import::from("weights", "wtd.cor")
```

```{r conflicts, include = FALSE}
tidyverse::tidyverse_conflicts()

conflicted::conflict_scout()
```

```{r knit-setup, include = FALSE}
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())

knitr::opts_chunk$set(cache = FALSE,
                      prompt = FALSE,
                      tidy = FALSE,
                      comment = NA,
                      message = FALSE,
                      warning = FALSE)

knitr::opts_knit$set(width = 75)
```

<!-- ----------------------------------------------------- -->

# Data 

## Spatial 

```{r}
kt = read_rds("data/BfS/kt.Rds")
gg = read_rds("data/BfS/gg.Rds")
tg3o = read_rds("data/BfS/tg3o.Rds")
se_alt = read_rds("data/BfS/se_alt.Rds")
carto = read_rds("data/carto/carto.Rds")
```

## Downscaled data

```{r}
exp_deaths_2020_year_gem = read_rds("results/exp_deaths_2020_year_gem.Rds") %>% 
  select(-munici_excess_pop, -munici_excess_rat) %>% 
  mutate(munici_sir = munici_observed / munici_expected_med)
```

```{r}
summary(exp_deaths_2020_year_gem$munici_sir)
```

Problematic cases 

```{r}
# exp_deaths_2020_year_gem %>% 
#   filter(is.na(munici_sir)) %>% 
#   select(munici_observed, munici_expected_med)

# zero and zero
exp_deaths_2020_year_gem %>% 
  filter(is.nan(munici_sir)) %>% 
  select(munici_observed, munici_expected_med, munici_sir)

# positive and zero
exp_deaths_2020_year_gem %>% 
  filter(is.infinite(munici_sir)) %>% 
  select(munici_observed, munici_expected_med, munici_sir)

# removed for now
exp_deaths_2020_year_gem %<>% 
  filter(!is.infinite(munici_sir)) %>% 
  filter(!is.nan(munici_sir))
```


<!-- ----------------------------------------------------- -->

# SIR

## Distribution 

```{r echo=FALSE}
ggplot(exp_deaths_2020_year_gem, aes(x = munici_sir)) + 
  geom_histogram(binwidth = 0.01) + 
  xlab("SIR") + ylab("Number of communities") + 
  theme_light()
```

## Maps

```{r include=FALSE}
tg3o %<>%
  select(GMDNR) %>% 
  right_join(exp_deaths_2020_year_gem) %>% 
  mutate(munici_sir_cat = Hmisc::cut2(munici_sir, g = 5))
```

```{r echo=FALSE}
frq(tg3o$munici_sir_cat)
```

### Choropleth 

```{r echo=FALSE}
tmap_mode("plot")

tm_shape(se_alt, is.master = FALSE) +
  tm_fill(col = "#a7cdf2") + 
  tm_shape(tg3o, is.master = TRUE) +
  tm_fill("munici_sir_cat", palette = "YlOrBr", 
          title = "SIR") + 
  tm_shape(kt) +
  tm_borders(col = "grey60")
```

### Proportional symbols 

Symbol size perceptually scaled to number of expected deaths.   

```{r echo=FALSE}
tm_shape(se_alt, is.master = FALSE) +
  tm_fill(col = "#a7cdf2") + 
  tm_shape(kt, is.master = TRUE) +
  tm_borders() +  
  tm_shape(tg3o) +
  tm_symbols(
    size = "munici_expected_med",
    # size = "munici_observed",
    # size = "munici_pop",
    col = "munici_sir_cat",
    shape = 15, scale = 2, perceptual = TRUE, 
    legend.size.show = TRUE, 
    palette = "YlOrBr", 
    title.size = "munici_pop", 
    title.col = "SIR") + 
  tm_legend(position = c("LEFT", "TOP"))
```

### Cartogram 

```{r include=FALSE}
carto_gg <- st_read("data/carto/out_gg.shp") %>% 
  st_set_crs(value = 2056) %>% 
  select(GMDNR) %>% 
  right_join(exp_deaths_2020_year_gem) %>% 
  mutate(munici_sir_cat = Hmisc::cut2(munici_sir, g = 5))

carto_se <- st_read("data/carto/out_se.shp") %>% 
  st_set_crs(value = 2056)

carto_kt <- st_read("data/carto/out_kt.shp") %>% 
  st_set_crs(value = 2056)

# frq(carto$munici_sir_cat)
```

```{r echo=FALSE}
carto = tm_shape(carto_se, is.master = FALSE) +
  tm_fill(col = "#a7cdf2") + 
  tm_shape(carto_gg, is.master = TRUE) +
  tm_fill("munici_sir_cat", palette = "YlOrBr", 
          title = "SIR") + 
  tm_shape(carto_kt) +
  tm_borders(col = "grey60")

carto
```

```{r}
tmap_save(carto, "docs/sir.png", width = 1920, height = 1080, asp = 0)
```

<!-- ----------------------------------------------------- -->

# EDA June vote 

## Map 

```{r echo=FALSE}
tmap_mode("view")

tm_shape(se_alt, is.master = FALSE) +
  tm_fill(col = "#a7cdf2") + 
  tm_shape(tg3o, is.master = TRUE) +
  tm_fill(c("munici_sir_cat", "vote_yes_jun_cat"), 
          palette = list("YlOrBr", "-YlOrBr"), 
          title = c("SIR", "% voting yes")) + 
  tm_shape(kt) +
  tm_borders(col = "grey60")
```
<br/>
<br/>

## Correlations {.tabset}

### Unweighted 

```{r}
cor.test(exp_deaths_2020_year_gem$munici_sir, 
         exp_deaths_2020_year_gem$vote_yes_jun_perc, 
         method = "pearson")
```

### Weighted

```{r}
wtd.cor(exp_deaths_2020_year_gem$munici_sir, 
        exp_deaths_2020_year_gem$vote_yes_jun_perc, 
        weight = exp_deaths_2020_year_gem$munici_expected_med)
```

## Scatter {.tabset}

### Unweighted

```{r echo=FALSE, warning=FALSE}
exp_deaths_2020_year_gem %>% 
  ggplot(aes(x = munici_sir, y = vote_yes_jun_perc)) + 
  geom_smooth(alpha = 0.25) +
  geom_point(aes(color = factor(vote_yes_jun_cat)), alpha = 0.2) + 
  scale_color_brewer(palette = "PuOr", direction = -1) + 
  theme_ipsum() +
  xlab("SIR") + ylab("% vote yes") 
```

### Weighted

```{r echo=FALSE, warning=FALSE}
exp_deaths_2020_year_gem %>% 
  ggplot(aes(x = munici_sir, y = vote_yes_jun_perc)) + 
  geom_smooth(aes(weight = munici_expected_med)) +
  geom_point(aes(color = factor(vote_yes_jun_cat), size = munici_expected_med), alpha = 0.2) + 
  scale_color_brewer(palette = "PuOr", direction = -1) + 
  theme_ipsum() +
  theme(legend.position = "none") +
  xlab("SIR") + ylab("% vote yes")
```

## Box 

```{r echo=FALSE}
ggplot(exp_deaths_2020_year_gem, aes(x = factor(vote_yes_jun_cat), y = munici_sir)) + 
  geom_boxplot() + 
  ylab("SIR") + xlab("% vote yes") + 
  theme_light()
```

<!-- ----------------------------------------------------- -->

# EDA Nov vote 

## Map 

```{r echo=FALSE}
tmap_mode("view")

tm_shape(se_alt, is.master = FALSE) +
  tm_fill(col = "#a7cdf2") + 
  tm_shape(tg3o, is.master = TRUE) +
  tm_fill(c("munici_sir_cat", "vote_yes_nov_cat"), 
          palette = list("YlOrBr", "-YlOrBr"), 
          title = c("SIR", "% voting yes")) + 
  tm_shape(kt) +
  tm_borders(col = "grey60")
```
<br/>
<br/>

## Correlations {.tabset}

### Unweighted 

```{r}
cor.test(exp_deaths_2020_year_gem$munici_sir, 
         exp_deaths_2020_year_gem$vote_yes_nov_perc, 
         method = "pearson")
```

### Weighted

```{r}
wtd.cor(exp_deaths_2020_year_gem$munici_sir, 
        exp_deaths_2020_year_gem$vote_yes_nov_perc, 
        weight = exp_deaths_2020_year_gem$munici_expected_med)
```

## Scatter {.tabset}

### Unweighted

```{r echo=FALSE, warning=FALSE}
exp_deaths_2020_year_gem %>% 
  ggplot(aes(x = munici_sir, y = vote_yes_nov_perc)) + 
  geom_smooth(alpha = 0.25) +
  geom_point(aes(color = factor(vote_yes_nov_cat)), alpha = 0.2) + 
  scale_color_brewer(palette = "PuOr", direction = -1) + 
  theme_ipsum() +
  xlab("SIR") + ylab("% vote yes") 
```

### Weighted

```{r echo=FALSE, warning=FALSE}
exp_deaths_2020_year_gem %>% 
  ggplot(aes(x = munici_sir, y = vote_yes_nov_perc)) + 
  geom_smooth(aes(weight = munici_expected_med)) +
  geom_point(aes(color = factor(vote_yes_nov_cat), size = munici_expected_med), alpha = 0.2) + 
  scale_color_brewer(palette = "PuOr", direction = -1) + 
  theme_ipsum() +
  theme(legend.position = "none") +
  xlab("SIR") + ylab("% vote yes")
```

## Box 

```{r echo=FALSE}
ggplot(exp_deaths_2020_year_gem, aes(x = factor(vote_yes_nov_cat), y = munici_sir)) + 
  geom_boxplot() + 
  ylab("SIR") + xlab("% vote yes") + 
  theme_light()
```
